requirements.txt
apispec==6.8.4
attrs==25.4.0
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
click==8.1.8
flasgger==0.9.7.1
Flask==3.1.2
Flask-JWT-Extended==4.7.1
flask-smorest==0.46.2
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==5.21.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
marshmallow==4.0.1
mistune==3.1.4
packaging==25.0
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
webargs==8.7.1
Werkzeug==3.1.3
zipp==3.23.0
